fix(channels): split setup adapter types
