+++ src/infra/approval-handler-runtime-types.ts
@@ -1,7 +1,7 @@
 // Defines channel-native approval handler runtime types.
 import type { OpenClawConfig } from "../config/types.openclaw.js";
 import type { ChannelApprovalNativePlannedTarget } from "./approval-native-delivery.js";
-import type { PreparedChannelNativeApprovalTarget } from "./approval-native-runtime.js";
+import type { PreparedChannelNativeApprovalTarget } from "./approval-native-runtime-types.js";
 import type { ChannelApprovalKind } from "./approval-types.js";
 import type {
   ExpiredApprovalView,
